+++ src/EnvironmentMonitor.Infrastructure/Data/Migrations/MeasurementDbContextModelSnapshot.cs
@@ -243,11 +243,11 @@ namespace EnvironmentMonitor.Infrastructure.Data.Migrations
 
             modelBuilder.Entity("EnvironmentMonitor.Domain.Entities.DeviceMessage", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
 
                     b.Property<int>("DeviceId")
                         .HasColumnType("int");
@@ -279,8 +279,8 @@ namespace EnvironmentMonitor.Infrastructure.Data.Migrations
                     b.Property<int>("DeviceId")
                         .HasColumnType("int");
 
-                    b.Property<int?>("DeviceMessageId")
-                        .HasColumnType("int");
+                    b.Property<long?>("DeviceMessageId")
+                        .HasColumnType("bigint");
 
                     b.Property<string>("Message")
                         .HasMaxLength(256)
@@ -406,8 +406,8 @@ namespace EnvironmentMonitor.Infrastructure.Data.Migrations
                         .HasColumnType("datetime2")
                         .HasDefaultValueSql("GETUTCDATE()");
 
-                    b.Property<int?>("DeviceMessageId")
-                        .HasColumnType("int");
+                    b.Property<long?>("DeviceMessageId")
+                        .HasColumnType("bigint");
 
                     b.Property<int>("SensorId")
                         .HasColumnType("int");
